fix rplchar
the checks if the space in the string buffer is sufficient for
the desired concatenation were wrong. This had the potential to
write beyond the end of the string.
If the end of the string is reached now a replacement might be incomplete
and the string is truncated to fit.